refactor: complete import path standardization across application
- Update all remaining relative imports to use @/ alias
- Replace @prisma/client imports with @generated/prisma path
- Fix admin-only guard role check logic
- Standardize type imports across guards, resolvers, and services
- Ensure consistent import patterns throughout the codebase
