+++ src/app.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ConfigModule, ConfigService } from '@nestjs/config';
+import { ConfigModule } from '@nestjs/config';
 import { GqlModule } from './gql/gql.module';
 import { UsersModule } from './users/users.module';
 import { AuthModule } from './auth/auth.module';
