+++ src/user-follows/user-follows.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
-import { Prisma, UserFollow } from '@prisma/client';
-import { PrismaService } from 'src/prisma/prisma.service';
+import { Prisma, UserFollow } from '@generated/prisma';
+import { PrismaService } from '@/prisma/prisma.service';
 
 @Injectable()
 export class UserFollowsService {
